services/user-dashboard-api/pyproject.toml
[tool.poetry]
name = "user-dashboard-api"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
google-cloud-bigquery = "^3.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
